Changed from using spheres for the kMT plus ends to point sets
One point set is used for each end of the spindle. This is much
faster to load and produces more accurate fluorescence images
because the fluorophores are located at the exact positions in
the simulation file whereas the single fluorophore on the sphere
was located anywhere on the surface of the 20nm diameter sphere
used to represent the point.

Added a function to produce the XML for one end of the spindle
to reduce code redundancy.